// src/app/shared/model/ScoreCategory/RefScoreCategoryTypeObj.model.ts
export class RefScoreCategoryTypeObj {
    RefScoreCategoryTypeId: number;
    RefScoreCategoryTypeCode: string;
    Descr: string;
    Notes: string;
    IsActive: boolean;
    ReservedField1: string;
    ReservedField2: string;
    ReservedField3: string;
    ReservedField4: string;
    ReservedField5: string;
    RowVersion: string;

    constructor() {
        this.RowVersion = "";
    }
}